Rename system package to amp
